refactoring joke search form, delete search view dependency from global store
