--- src/components/CategoryBtn/index.js
@@ -1,3 +1,0 @@
-import CategoryBtn from "./CategoryBtn";
-
-export default CategoryBtn;
