+++ src/components/JokeSearchForm/JokeSearchForm.js
@@ -1,40 +1,16 @@
-import React, { useEffect } from "react";
+import React, { useEffect, useState } from "react";
 
-import { useDispatch, useSelector } from "react-redux";
-import {
-  changeJokeSearchType,
-  changeSelectedJokeCategory,
-  setSearchQuery,
-} from "../../store/actions/jokeSearch";
+import { useDispatch } from "react-redux";
+import { changeJokeSearchType } from "../../store/actions/jokeSearch";
 import { fetchJokeCategories } from "../../store/thunks/fetchCategoriesThunk";
-import CategoryBtn from "../CategoryBtn";
 
-const selectJokeSearch = (state) => state.jokeSearch;
+import CategoriesBtns from "./components/CategoriesBtns";
+import TextSearchInput from "./components/TextSearchInput";
 
 export default function JokeSearchFormComponent({ handleGetJoke }) {
-  const { searchType: jokeSearchType, categories: jokeSearchCategories } =
-    useSelector(selectJokeSearch);
+  const [searchView, setSearchView] = useState(null);
 
   const dispatch = useDispatch();
-
-  //elements
-  const jokeSearchCategoriesBtns = jokeSearchCategories.map(
-    (category, index) => {
-      const categoryBtnProps = {
-        key: index,
-        category: category,
-        onClick: handleChangeSelectedJokeCategory,
-      };
-
-      return <CategoryBtn {...categoryBtnProps} />;
-    }
-  );
-
-  //styles
-  const categoriesBtnsDisplay = jokeSearchType === "category" ? "" : "none";
-  const categoriesBtnsStyles = { display: categoriesBtnsDisplay };
-
-  const textSearchInputDisplay = jokeSearchType === "search" ? "block" : "";
 
   //props
   const randomRadioInputProps = {
@@ -61,13 +37,6 @@ export default function JokeSearchFormComponent({ handleGetJoke }) {
     id: "searchRadioInput",
   };
 
-  const textSearchInputProps = {
-    type: "text",
-    placeholder: "Free text search... ",
-    style: { display: textSearchInputDisplay },
-    onChange: handleChangeSearchQuery,
-  };
-
   //effects
   useEffect(() => {
     dispatch(fetchJokeCategories);
@@ -77,31 +46,8 @@ export default function JokeSearchFormComponent({ handleGetJoke }) {
     const value = e.target.value;
     const action = changeJokeSearchType(value);
     dispatch(action);
-  }
 
-  function handleChangeSelectedJokeCategory(e) {
-    const value = e.target.value;
-    const action = changeSelectedJokeCategory(value);
-    dispatch(action);
-  }
-
-  function handleChangeSearchQuery(e) {
-    const value = e.target.value;
-    const style = e.target.style;
-    // console.dir(e.target);
-    const maxQueryLength = 120;
-    const minQueryLength = 3;
-
-    if (value.length < minQueryLength || value.length > maxQueryLength) {
-      e.target.placeholder = "Request must be more than 3 characters.";
-      style.borderColor = "red";
-    }
-
-    if (value.length >= minQueryLength && value.length <= maxQueryLength) {
-      style.borderColor = "";
-      const action = setSearchQuery(value);
-      dispatch(action);
-    }
+    setSearchView(value);
   }
 
   return (
@@ -114,14 +60,12 @@ export default function JokeSearchFormComponent({ handleGetJoke }) {
         <input {...categoriesRadioInputProps} />
         <label htmlFor="categoriesRadioInput">Сategories</label>
       </div>
-      <div className="categories-btns" style={categoriesBtnsStyles}>
-        {jokeSearchCategoriesBtns}
-      </div>
+      <CategoriesBtns searchView={searchView} />
       <div>
         <input {...searchRadioInputProps} />
         <label htmlFor="searchRadioInput">Search</label>
       </div>
-      <input {...textSearchInputProps} />
+      <TextSearchInput searchView={searchView} />
       <button type="submit">Get a joke</button>
     </form>
   );
